Extract results

+++ back/app/drools/recommendation/Recommendation.java
@@ -1,14 +1,15 @@
 package drools.recommendation;
 
-import java.util.Map;
+
+
 
 public class Recommendation {
 
-    private Long recipeId;
+    private int recipeId;
 
     private Double hit;
 
-    public Recommendation(Long recipeId, Double hit) {
+    public Recommendation(int recipeId, Double hit) {
         this.recipeId = recipeId;
         this.hit = hit;
     }
